adminbot/pyproject.toml
[project]
name = "dndadminbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram==3.22.0",
    "requests==2.32.5",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiogram-dialog==2.4.0",
    "redis (>=6.4.0,<7.0.0)",
    "httpx==0.28.1",
    "pillow==11.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.0",
    "coverage==7.10.7",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
builtins = []
cache-dir = ".ruff_cache"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "dist",
    "migrations",
    "node_modules",
    "venv",
    "manage.py"
]
extend-exclude = []
extend-include = []
fix = false
fix-only = false
force-exclude = true
include = ["*.py", "*.pyi", "*.ipynb", "**/pyproject.toml"]
indent-width = 4
line-length = 79

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120
exclude = []
indent-style = "space"
line-ending = "lf"
preview = false
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
exclude = ["tests.py"]
fixable = ["ALL"]
ignore = [
    "ARG",
    "COM812",
    "D",
    "ISC001",
    "PT009",
    "N813",
    "RUF012",
    "ANN201",
    "RUF001",
]
select = ["ALL"]
task-tags = ["TODO"]
